Added walls to the map, enemies no longer get off screen and instead are bouncing on the walls * Additionally switched Uint8 for Uint16

+++ 2DGameEngine/2DGameEngine/src/Components/SpriteComponent.h
@@ -9,8 +9,8 @@
 	\param unsigned int height = 32
 	\param  unsigned int zIndex = 0
 	\param bool isCameraIndependent = false
-	\param Uint8 tilemapXCoord = 0
-	\param Uint8 tilemapYCoord = 0
+	\param Uint16 tilemapXCoord = 0
+	\param Uint16 tilemapYCoord = 0
 */
 struct SpriteComponent
 {
@@ -24,7 +24,7 @@ struct SpriteComponent
 	SDL_Rect m_textureRect;
 	
 
-	SpriteComponent(const std::string& assetId = "", unsigned int width = 32, unsigned int height = 32, unsigned int zIndex = 0, bool isCameraIndependent = false, Uint8 tilemapXCoord = 0, Uint8 tilemapYCoord = 0)
+	SpriteComponent(const std::string& assetId = "", unsigned int width = 32, unsigned int height = 32, unsigned int zIndex = 0, bool isCameraIndependent = false, Uint16 tilemapXCoord = 0, Uint16 tilemapYCoord = 0)
 		: m_assetId(assetId)
 		, m_width(width)
 		, m_height(height)
